Refactor qchem ts_optimizer and add qchem optimizer
